refactor: Line 및 Ladder 생성 로직 분리
