+++ src/main/java/nextstep/ladder/model/Ladder.java
@@ -1,23 +1,13 @@
 package nextstep.ladder.model;
 
-import static java.util.stream.Collectors.*;
-
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.IntStream;
-
-import nextstep.ladder.model.strategy.PointPickerStrategy;
 
 public class Ladder {
-    private static final String INVALID_LADDER_LENGTH_EXCEPTION_MESSAGE = "사다리의 높이는 양수여야 합니다.";
-
     private final List<Line> lines;
 
-    public Ladder(int ladderLength, int countOfPerson, PointPickerStrategy strategy) {
-        validate(ladderLength);
-        this.lines = IntStream.range(0, ladderLength)
-            .mapToObj(idx -> new Line(countOfPerson, strategy))
-            .collect(toList());
+    public Ladder(List<Line> lines) {
+        this.lines = lines;
     }
 
     int move(int startPosition) {
@@ -32,11 +22,5 @@ public class Ladder {
 
     public List<Line> getLines() {
         return Collections.unmodifiableList(lines);
-    }
-
-    private void validate(int ladderLength) {
-        if (ladderLength <= 0) {
-            throw new IllegalArgumentException(INVALID_LADDER_LENGTH_EXCEPTION_MESSAGE);
-        }
     }
 }
