+++ src/test/java/nextstep/ladder/model/LadderTest.java
@@ -9,32 +9,15 @@ import java.util.stream.Stream;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
+
+import nextstep.ladder.model.factory.LadderFactory;
 
 class LadderTest {
-    @ParameterizedTest(name = "사다리의 길이와 사람 수가 주어질 때, 사다리를 생성한다. 길이: {0}, 사람 수: {1}")
-    @CsvSource({
-        "3,3",
-        "1,3"
-    })
-    void create(int ladderLength, int countOfPerson) {
-        new Ladder(ladderLength, countOfPerson, () -> true);
-    }
-
-    @ParameterizedTest(name = "부적절한 사다리의 길이가 주어졌을 때, 예외가 던져진다. 길이: {0}")
-    @ValueSource(ints = {-1, 0})
-    void createWithInvalidLength(int ladderLength) {
-        assertThatThrownBy(() -> new Ladder(ladderLength, 3, () -> true))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessageContaining("사다리의 높이는 양수여야 합니다");
-    }
-
     @ParameterizedTest(name = "사다리 길이와 사람 수가 주어졌을 때, 각 위치에 대하여 사다리 타기를 수행합니다; 사다리 길이: {0}, 사람 수: {1}")
     @MethodSource("provideLadderMovingSource")
     void move(int ladderLength, int countOfPerson, List<Integer> expectedPositions) {
-        Ladder ladder = new Ladder(ladderLength, countOfPerson, () -> true);
+        Ladder ladder = LadderFactory.create(ladderLength, countOfPerson, () -> true);
 
         List<Integer> positions = IntStream.range(0, countOfPerson)
             .map(ladder::move)
